AI_DESKTOP_ASSISTANT/requirements.txt
# ===============================
# UI (PyQt)
# ===============================
PyQt5==5.15.10
PyQt5-sip>=12.11

# ===============================
# AI / API
# ===============================
requests>=2.31.0
groq>=0.9.0

# ===============================
# SPEECH / VOICE
# ===============================
SpeechRecognition>=3.10.0
PyAudio>=0.2.13
pyperclip>=1.8.2

# ===============================
# OCR & SCREEN CAPTURE
# ===============================
pytesseract==0.3.10
Pillow>=9.5.0
pyautogui>=0.9.54

# ===============================
# HOTKEYS / THREADING
# ===============================
keyboard>=0.13.5

# ===============================
# WINDOWS / STEALTH / WIN32
# ===============================
pywin32>=306
psutil>=5.9.6

# ===============================
# BUILD (EXE)
# ===============================
pyinstaller>=6.5.0
